requirements.txt
absl-py==2.1.0
albumentations==1.3.0
appdirs==1.4.4
cachetools==5.3.2
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
cmake==3.25.0
contourpy==1.0.7
cuda-python==12.1.0
cycler==0.11.0
Cython==0.29.35
docker-pycreds==0.4.0
filelock==3.9.0
fonttools==4.39.4
fsspec==2023.5.0
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.26.2
google-auth-oauthlib==1.0.0
grpcio==1.60.0
huggingface-hub==0.15.1
idna==3.4
imageio==2.30.0
importlib-metadata==7.0.1
importlib-resources==5.12.0
Jinja2==3.1.2
joblib==1.2.0
jsonpatch==1.32
jsonpointer==2.3
kiwisolver==1.4.4
lazy_loader==0.2
lit==15.0.7
Markdown==3.5.2
MarkupSafe==2.1.2
matplotlib==3.7.1
mpmath==1.2.1
networkx==3.0
numpy==1.23.5
oauthlib==3.2.2
opencv-contrib-python==4.7.0.72
opencv-python==4.7.0.72
opencv-python-headless==4.7.0.72
packaging==23.1
pandas==2.0.2
pathtools==0.1.2
Pillow==9.3.0
protobuf==4.23.2
psutil==5.9.5
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0
qudida==0.0.4
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
safetensors==0.3.1
scikit-image==0.21.0
scikit-learn==1.2.2
scipy==1.10.1
sentry-sdk==1.25.0
setproctitle==1.3.2
six==1.16.0
sk-video==1.1.10
smmap==5.0.0
sympy==1.11.1
tensorboard==2.14.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
termcolor==2.3.0
threadpoolctl==3.1.0
tifffile==2023.4.12
timm==0.9.2
torch==2.0.1+cu117
torch-tb-profiler==0.4.3
torchaudio==2.0.2+cu117
torchvision==0.15.2+cu117
tornado==6.3.2
tqdm==4.65.0
triton==2.0.0
typing_extensions==4.4.0
tzdata==2023.3
urllib3==1.26.13
visdom==0.2.4
wandb==0.15.3
websocket-client==1.5.2
Werkzeug==3.0.1
yacs==0.1.8
zipp==3.15.0
